Fix tx noti etherscan link
