+++ src/components/Navbar/TransactionList/index.ts
@@ -1,0 +1,1 @@
+export { TransactionsList } from './TransactionsList'
