+++ src/utils/etherscan.ts
@@ -1,7 +1,11 @@
 import { readNetwork } from 'constants/networks'
 import { NetworkName } from 'models/network-name'
 
-export const etherscanLinkForTxHash = (hash: string) =>
-  `https://${readNetwork.name}${
-    readNetwork.name !== NetworkName.mainnet ? '.' : ''
-  }etherscan.io/tx/${hash}`
+export const etherscanLink = (type: 'tx' | 'address', hash: string) => {
+  let subdomain = ''
+  if (readNetwork.name !== NetworkName.mainnet) {
+    subdomain = readNetwork.name + '.'
+  }
+
+  return `https://${subdomain}etherscan.io/${type}/${hash}`
+}
